refactor : change name files and reorganice tests
