+++ Cargo.toml
@@ -7,6 +7,7 @@ repository = "https://github.com/machinalabs/earthmind-rs.git "
 
 [lib]
 crate-type = ["cdylib", "rlib"]
+path = "./src/contract.rs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
